Added support for .NET Platform 5.4, .NET Framework 4.6, and .NET Framework 4.6.1. Produced a new nuget package, 1.0.1-beta-3
