+++ Crc32.sln
@@ -1,44 +1,14 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 2013
-VisualStudioVersion = 12.0.30501.0
+# Visual Studio 14
+VisualStudioVersion = 14.0.24720.0
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Net35", "Net35", "{9AEC4BB1-A8EF-4CBA-9DAC-7D26F4CE0357}"
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Net40", "Net40", "{99221B8F-DD7C-4A8C-8729-4DC7A0AA23EB}"
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Net45", "Net45", "{B694F6BE-6E33-4656-80D7-BBCF354001B1}"
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Net451", "Net451", "{79AE5203-31D6-46F9-86C1-D966223F81EE}"
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Net452", "Net452", "{78A8A146-B8AA-4577-AACA-85002EFF0522}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{4F4AAC76-FA2B-4B22-BDB8-CFCEF90FF71C}"
 	ProjectSection(SolutionItems) = preProject
 		src\Properties\AssemblyInfo.cs = src\Properties\AssemblyInfo.cs
-		src\Crc32Algorithm.cs = src\Crc32Algorithm.cs
 	EndProjectSection
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Net20", "Net20", "{E273D6EA-D252-4205-8F84-92EC22F33BE5}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Crc32", "Net20\Crc32.csproj", "{E18C3E10-2820-4CE0-8E10-4E5C0F810914}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Crc32", "Net35\Crc32.csproj", "{FDC6129D-C934-4831-A7A1-219DFE56F01D}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Crc32", "Net40\Crc32.csproj", "{532493E1-A7B3-454A-BD28-32075970F5FD}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Crc32", "Net45\Crc32.csproj", "{EC96FB52-AB9A-403D-9CC0-29976576BE13}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Crc32", "Net451\Crc32.csproj", "{66E44A83-BC4E-47AF-81F7-A2D2C2D24CFA}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Crc32", "Net452\Crc32.csproj", "{F667368F-90DA-4A37-B01F-24FDF69E914A}"
-	ProjectSection(ProjectDependencies) = postProject
-		{E18C3E10-2820-4CE0-8E10-4E5C0F810914} = {E18C3E10-2820-4CE0-8E10-4E5C0F810914}
-		{EC96FB52-AB9A-403D-9CC0-29976576BE13} = {EC96FB52-AB9A-403D-9CC0-29976576BE13}
-		{66E44A83-BC4E-47AF-81F7-A2D2C2D24CFA} = {66E44A83-BC4E-47AF-81F7-A2D2C2D24CFA}
-		{FDC6129D-C934-4831-A7A1-219DFE56F01D} = {FDC6129D-C934-4831-A7A1-219DFE56F01D}
-		{532493E1-A7B3-454A-BD28-32075970F5FD} = {532493E1-A7B3-454A-BD28-32075970F5FD}
-	EndProjectSection
+Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "Crc32", "src\Crc32\Crc32.xproj", "{934EC66A-D19D-4E03-A260-A3230576F8A6}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -46,40 +16,15 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{E18C3E10-2820-4CE0-8E10-4E5C0F810914}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{E18C3E10-2820-4CE0-8E10-4E5C0F810914}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{E18C3E10-2820-4CE0-8E10-4E5C0F810914}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{E18C3E10-2820-4CE0-8E10-4E5C0F810914}.Release|Any CPU.Build.0 = Release|Any CPU
-		{FDC6129D-C934-4831-A7A1-219DFE56F01D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{FDC6129D-C934-4831-A7A1-219DFE56F01D}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{FDC6129D-C934-4831-A7A1-219DFE56F01D}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{FDC6129D-C934-4831-A7A1-219DFE56F01D}.Release|Any CPU.Build.0 = Release|Any CPU
-		{532493E1-A7B3-454A-BD28-32075970F5FD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{532493E1-A7B3-454A-BD28-32075970F5FD}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{532493E1-A7B3-454A-BD28-32075970F5FD}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{532493E1-A7B3-454A-BD28-32075970F5FD}.Release|Any CPU.Build.0 = Release|Any CPU
-		{EC96FB52-AB9A-403D-9CC0-29976576BE13}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{EC96FB52-AB9A-403D-9CC0-29976576BE13}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{EC96FB52-AB9A-403D-9CC0-29976576BE13}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{EC96FB52-AB9A-403D-9CC0-29976576BE13}.Release|Any CPU.Build.0 = Release|Any CPU
-		{66E44A83-BC4E-47AF-81F7-A2D2C2D24CFA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{66E44A83-BC4E-47AF-81F7-A2D2C2D24CFA}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{66E44A83-BC4E-47AF-81F7-A2D2C2D24CFA}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{66E44A83-BC4E-47AF-81F7-A2D2C2D24CFA}.Release|Any CPU.Build.0 = Release|Any CPU
-		{F667368F-90DA-4A37-B01F-24FDF69E914A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{F667368F-90DA-4A37-B01F-24FDF69E914A}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{F667368F-90DA-4A37-B01F-24FDF69E914A}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{F667368F-90DA-4A37-B01F-24FDF69E914A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{934EC66A-D19D-4E03-A260-A3230576F8A6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{934EC66A-D19D-4E03-A260-A3230576F8A6}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{934EC66A-D19D-4E03-A260-A3230576F8A6}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{934EC66A-D19D-4E03-A260-A3230576F8A6}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
-		{E18C3E10-2820-4CE0-8E10-4E5C0F810914} = {E273D6EA-D252-4205-8F84-92EC22F33BE5}
-		{FDC6129D-C934-4831-A7A1-219DFE56F01D} = {9AEC4BB1-A8EF-4CBA-9DAC-7D26F4CE0357}
-		{532493E1-A7B3-454A-BD28-32075970F5FD} = {99221B8F-DD7C-4A8C-8729-4DC7A0AA23EB}
-		{EC96FB52-AB9A-403D-9CC0-29976576BE13} = {B694F6BE-6E33-4656-80D7-BBCF354001B1}
-		{66E44A83-BC4E-47AF-81F7-A2D2C2D24CFA} = {79AE5203-31D6-46F9-86C1-D966223F81EE}
-		{F667368F-90DA-4A37-B01F-24FDF69E914A} = {78A8A146-B8AA-4577-AACA-85002EFF0522}
+		{934EC66A-D19D-4E03-A260-A3230576F8A6} = {4F4AAC76-FA2B-4B22-BDB8-CFCEF90FF71C}
 	EndGlobalSection
 EndGlobal
